+++ TabpediaFin/TabpediaFin/Handler/ContactPersonHandler/ContactPersonDeleteHandler.cs
@@ -1,30 +1,36 @@
 ﻿namespace TabpediaFin.Handler.ContactPersonHandler
 {
-    public class ContactPersonDeleteHandler : IRequestHandler<ContactPersonDeleteDto, RowResponse<bool>>
+    public class ContactPersonDeleteHandler : IDeleteByIdHandler<ContactPersonFetchDto>
     {
         private readonly FinContext _context;
-        private readonly ICurrentUser _currentUser;
+        private readonly IContactPersonCacheRemover _cacheRemover;
 
-        public ContactPersonDeleteHandler(FinContext db, ICurrentUser currentUser)
+        public ContactPersonDeleteHandler(FinContext db, IContactPersonCacheRemover cacheRemover)
         {
             _context = db;
-            _currentUser = currentUser;
+            _cacheRemover = cacheRemover;
         }
-        public async Task<RowResponse<bool>> Handle(ContactPersonDeleteDto request, CancellationToken cancellationToken)
+
+        public async Task<RowResponse<ContactPersonFetchDto>> Handle(DeleteByIdRequestDto<ContactPersonFetchDto> request, CancellationToken cancellationToken)
         {
-            var result = new RowResponse<bool>();
+            var result = new RowResponse<ContactPersonFetchDto>();
+
             try
             {
-                var ContactPerson = await _context.ContactPerson.FirstAsync(x => x.Id == request.Id && x.TenantId == _currentUser.TenantId, cancellationToken);
+                var ContactPerson = await _context.ContactPerson.FirstAsync(x => x.Id == request.Id, cancellationToken);
+                if (ContactPerson == null || ContactPerson.Id == 0)
+                {
+                    throw new HttpException(HttpStatusCode.NotFound, "Data not found");
+                }
 
-                _context.ContactPerson.Attach(ContactPerson);
                 _context.ContactPerson.Remove(ContactPerson);
-
                 await _context.SaveChangesAsync(cancellationToken);
+
+                _cacheRemover.RemoveCache();
 
                 result.IsOk = true;
                 result.ErrorMessage = string.Empty;
-                result.Row = true;
+                result.Row = new ContactPersonFetchDto();
             }
             catch (Exception ex)
             {
@@ -34,11 +40,5 @@
 
             return result;
         }
-    }
-
-    public class ContactPersonDeleteDto : IRequest<RowResponse<bool>>
-    {
-        public int Id { get; set; }
-
     }
 }
